Stop marking servers as down
According to nginx upstream docs:

> down
>    marks the server as permanently unavailable. 

I suppose that this may be useful for users who are manually editing their nginx configuration file:
They can mark a server down without loosing the configuration. Is there any other purpose ?
For users who generate the configuration: just put the working servers in the configuration and skip
others.

The most disturbing for me is:

    server 127.0.0.1 down;

What is 127.0.0.1 ? It's the nginx-proxy container. 
Is it supposed to be an upstream ? Not in any case that I can think of.
What does having such line cause ? Well I've had:

   [emerg] 28#28: invalid number of arguments in "upstream" directive

But also:

    Generated '/etc/nginx/conf.d/default.conf' from 12 containers
    Running 'nginx -s reload'
    Error running notify command: nginx -s reload, exit status 1
    Received event start for container 5c6cb0bf8e05

Yep, the whole LB is down and no helpful error message about what is going on.

I beleive this fixes what most users have been facing in #1132
As well as the regression introduced in #1106.

> Deleted code is debugged code.
        — Jeff Sickel